build: Require xcomposite
The X11TexturePixmap actor uses XComposite API directly, without guards.
It has been doing so for a while, against the fact that we do check for
the XComposite extension - but we don't depend on it. As soon as you try
building Clutter on X11 without the XComposite extension available all
hell breaks loose.

The obvious fix is to make Clutter depend on XComposite - basically
ratifying what's the current state of things.